refactor: move popover to a separate folder
rearrange and rename utilities and hooks related to a popover
