+++ src/core/packages/chrome/navigation/src/components/popover/use_popover_open.ts
@@ -7,7 +7,7 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import { useCallback, useEffect, useState } from 'react';
+import { useState, useCallback } from 'react';
 
 /**
  * Hook for managing popover open state
@@ -20,21 +20,4 @@ export const usePopoverOpen = (initialState = false) => {
   const toggle = useCallback(() => setIsOpen((prev) => !prev), []);
 
   return { isOpen, open, close, toggle };
-};
-
-/**
- * Hook for side panel effects
- * This hook is used to handle the popover state when the side panel is open.
- */
-export const useSidePanelEffect = (
-  isSidePanelOpen: boolean,
-  onClose: () => void,
-  clearTimeout: () => void
-) => {
-  useEffect(() => {
-    if (isSidePanelOpen) {
-      clearTimeout();
-      onClose();
-    }
-  }, [isSidePanelOpen, onClose, clearTimeout]);
 };
